remove useless style & rename
