+++ src/components/ScreenCapture.vue
